+++ app/src/main/java/com/github/polypoly/app/ui/menu/MenuComposable.kt
@@ -23,6 +23,7 @@ import androidx.core.content.ContextCompat.startActivity
 import com.github.polypoly.app.R
 import com.github.polypoly.app.base.RulesObject
 import com.github.polypoly.app.ui.menu.profile.ProfileActivity
+import com.github.polypoly.app.ui.menu.rankings.RankingsActivity
 import com.github.polypoly.app.ui.menu.settings.SettingsActivity
 
 object MenuComposable {
